Remove update.sh script, and cleanup install/remove scripts.

+++ remove.sh
@@ -4,6 +4,11 @@
 # Remove the post-installation configuration
 #
 
+# If there's no backup, don't do anything
+if [ ! -d backup ]; then
+	exit
+fi
+
 # Delete local configuration files
 for FILE in $(find config -type f | sed 's:^config::'); do
 	rm -fv $FILE
@@ -12,6 +17,5 @@ done
 # Restore previous configuration files
 cp -pRfv backup/* /
 
-# Archive the backup directory and remove it
-tar cvfJ backup-$(date "+%Y%m%d").tar.xz backup
+# Remove the backup folder
 rm -rfv backup
